+++ examples/twitter-clone/src/test/scala/finatra/quickstart/TwitterCloneStartupTest.scala
@@ -9,7 +9,7 @@ class TwitterCloneStartupTest extends FeatureTest {
   override val server = new EmbeddedHttpServer(
     stage = Stage.PRODUCTION,
     twitterServer = new TwitterCloneServer,
-    clientFlags = Map(
+    flags = Map(
       "com.twitter.server.resolverMap" -> "firebase=nil!"))
 
   "server" in {