+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/tweetexample/test/TweetsControllerIntegrationTest.scala
@@ -14,7 +14,7 @@ class TweetsControllerIntegrationTest extends FeatureTest {
     new TweetsEndpointServer,
     defaultRequestHeaders = Map("X-UserId" -> "123"),
     // Set client flags to also start on HTTPS port
-    clientFlags = Map(
+    flags = Map(
       "https.port" -> ":0",
       "cert.path" -> "",
       "key.path" -> ""))