+++ inject/inject-app/src/test/scala/com/twitter/inject/app/tests/SampleApp.scala
@@ -2,13 +2,14 @@ package com.twitter.inject.app.tests
 
 import com.twitter.inject.app.App
 
-object SampleGuiceAppMain extends SampleGuiceApp
+object SampleAppMain extends SampleApp
 
-class SampleGuiceApp extends App {
+class SampleApp extends App {
+  override val name = "sample-app"
 
   override val modules = Seq()
 
-  override def appMain() {
+  override def appMain(): Unit = {
     injector.instance[SampleManager].start()
   }
 }