+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/integration/EchoHttpServerFeatureTest.scala
@@ -13,7 +13,7 @@ class EchoHttpServerFeatureTest extends HttpTest {
 
   val httpServer = new EmbeddedHttpServer(
     twitterServer = new EchoHttpServer,
-    extraArgs = Seq(
+    args = Seq(
       "-thrift.clientId=echo-http-service",
       resolverMap("thrift-echo-service" -> thriftServer.thriftHostAndPort)))